+++ package.json
@@ -7,7 +7,8 @@
     "start": "node ./dist/server.js",
     "dev": "nodemon --watch 'src/**/*' -e ts,tsx --exec ts-node ./src/server.ts",
     "lint": "yarn run eslint . --ext=.ts",
-    "test": "yarn run jest",
+    "test:integration": "yarn run jest ./test/integration",
+    "test:unit": "yarn run jest ./test/unit",
     "migration:run": "yarn run ts-node ./node_modules/typeorm/cli.js migration:run",
     "migration:revert": "yarn run ts-node ./node_modules/typeorm/cli.js migration:revert",
     "migration:generate": "yarn run ts-node ./node_modules/typeorm/cli.js migration:generate"
